+++ tests/src/Kernel/Routing/RouteSubscriber/RouteSubscriberTest.php
@@ -25,7 +25,7 @@ class RouteSubscriberTest extends KernelTestBase {
   /**
    * {@inheritdoc}}
    */
-  protected function setUp() {
+  protected function setUp(): void {
     parent::setUp();
     \Drupal::service('router.builder')->rebuild();
   }